+++ lib/index.ts
@@ -1,11 +1,4 @@
 /*jslint node: true */
 'use strict';
-const gd = require('./get-diffs');
-
-module.exports = {
-  getDiffs: gd.getDiffs
-
-
-};
 
 export * from './get-diffs'